+++ src/app/interfaces/condition.ts
@@ -1,0 +1,6 @@
+export interface Condition {
+  condition: string;
+  label: string;
+  iconName: string;
+  active: boolean;
+}