+++ src/app/services/class.service.ts
@@ -1,10 +1,10 @@
-import { Injectable }               from '@angular/core';
+import { Injectable } from '@angular/core';
 import {
   AngularFireDatabase,
   AngularFireList,
   AngularFireObject,
-} from 'angularfire2/database';
-import { Character }                from '../character';
+}                     from 'angularfire2/database';
+import { Character }  from '../interfaces/character';
 
 @Injectable({
   providedIn: 'root'